pyproject.toml
[project]
name = "outlook-category-validator"
version = "1.0.0"
description = "Validator for OutlookCategory.yaml JSON Schema (Draft 2020-12)"
requires-python = ">=3.8"

authors = [
  { name = "Takashi Sasaki", email = "takashi316@gmail.com" }
]

dependencies = [
  "jsonschema[format]>=4.18.0",
  "PyYAML>=6.0",
  "doit>=0.36.0"
]

[tool.poetry]
package-mode = false

[tool.doit]
backend = "json"
